move the existing example to with-script-in-browser
