+++ packages/excalidraw/examples/with-script-in-browser/App.scss
