+++ packages/excalidraw/examples/with-script-in-browser/MobileFooter.tsx
@@ -1,4 +1,4 @@
-import type { ExcalidrawImperativeAPI } from "../types";
+import type { ExcalidrawImperativeAPI } from "../../types";
 import CustomFooter from "./CustomFooter";
 const { useDevice, Footer } = window.ExcalidrawLib;
