+++ packages/excalidraw/examples/with-script-in-browser/index.tsx
