+++ packages/excalidraw/examples/with-script-in-browser/initialData.tsx
@@ -1,5 +1,5 @@
-import type { ExcalidrawElementSkeleton } from "../data/transform";
-import type { FileId } from "../element/types";
+import type { ExcalidrawElementSkeleton } from "../../data/transform";
+import type { FileId } from "../../element/types";
 
 const elements: ExcalidrawElementSkeleton[] = [
   {
