+++ packages/excalidraw/examples/with-script-in-browser/public/images/excalibot.png
