+++ packages/excalidraw/examples/with-script-in-browser/public/images/pika.jpeg
