+++ packages/excalidraw/examples/with-script-in-browser/public/images/rocket.jpeg
