+++ packages/excalidraw/examples/with-script-in-browser/public/index.html
