+++ packages/excalidraw/examples/with-script-in-browser/sidebar/ExampleSidebar.tsx
